+++ lib/src/types.dart
@@ -1,3 +1,3 @@
 typedef JsonClassBuilder<T> = T Function(dynamic map);
 
-typedef JsonClassListBuilder<T> = List<T> Function(dynamic list);
+typedef JsonClassListBuilder<T> = List<T> Function(Iterable<dynamic> list);